[§] interface consistency
