+++ inference-engine/src/snippets/include/snippets/op/tile.hpp
@@ -1,0 +1,36 @@
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include <transformations_visibility.hpp>
+
+#include "ngraph/op/op.hpp"
+#include "snippets/emitter.hpp"
+
+namespace ngraph {
+namespace snippets {
+namespace op {
+
+/**
+ * @interface Tile
+ * @brief Generated by Canonicalization and represents Loop in affine notation
+ * @ingroup snippets
+ */
+class TRANSFORMATIONS_API Tile : public ngraph::op::Op {
+public:
+    NGRAPH_RTTI_DECLARATION;
+
+    Tile(const std::vector<std::pair<std::shared_ptr<ngraph::snippets::Emitter>, ngraph::snippets::RegInfo>>& region);
+    Tile() = default;
+    std::vector<std::pair<std::shared_ptr<ngraph::snippets::Emitter>, ngraph::snippets::RegInfo>> region;
+
+    std::shared_ptr<Node> clone_with_new_inputs(const OutputVector& inputs) const override {
+        return std::make_shared<Tile>(region);
+    }
+};
+
+} // namespace op
+} // namespace snippets
+} // namespace ngraph
