+++ inference-engine/src/snippets/src/op/nop.cpp
@@ -3,7 +3,6 @@
 //
 
 #include "snippets/op/nop.hpp"
-#include "snippets/generator.hpp"
 
 using namespace std;
 using namespace ngraph;
@@ -16,14 +15,4 @@ snippets::op::Nop::Nop(const OutputVector& arguments, const OutputVector& result
     x.insert(x.end(), results.begin(), results.end());
     return x;
     }()) {
-}
-
-NGRAPH_RTTI_DEFINITION(snippets::op::Tile, "Tile", 0);
-
-snippets::op::Tile::Tile(const std::vector<std::pair<std::shared_ptr<snippets::Emitter>, snippets::RegInfo>>& nested) : Op(), region(nested) {
-}
-
-NGRAPH_RTTI_DEFINITION(snippets::op::Kernel, "Kernel", 0);
-
-snippets::op::Kernel::Kernel(const std::vector<std::pair<std::shared_ptr<snippets::Emitter>, snippets::RegInfo>>& nested) : Op(), region(nested) {
 }
