+++ inference-engine/src/snippets/src/op/subgraph.cpp
@@ -12,6 +12,7 @@
 #include "snippets/pass/assign_registers.hpp"
 
 #include <ngraph/pass/manager.hpp>
+#include <transformations/serialize.hpp>
 
 #include <algorithm>
 #include <memory>
@@ -343,4 +344,13 @@ void snippets::op::Subgraph::print_statistics(bool verbose) {
     if (verbose) {
         this->print();
     }
+}
+
+void snippets::op::Subgraph::serialize() const {
+    std::stringstream xmlFile, binFile;
+    ngraph::pass::Serialize serializer(xmlFile, xmlFile, ngraph::pass::Serialize::Version::IR_V10);
+    serializer.run_on_function(get_body());
+    auto m_constants = binFile.str();
+    auto m_model = xmlFile.str();
+    std::cout << m_model << std::endl;
 }
